+++ book_data_loader.rb
